Create Matrix.cpp
This is the functions for the Matrix class, duh. I never made a new file on my personal computer...git whatever wasn't having it, but this should work fine. Maybe?! Who wants to play guess my output!!! (There is no actual output so i win, sucks) 
Also it's late and this is probably annoying to read but one more thing simpletons, talking to myself, but check deleteRow(int r) because i wasn't sure how to delete just the pointer in the matrix and I think I'm leaking more than a leaky faucet in Ruwanda on a hot summers day