Getting closer
All comments have been added and all functions implemented besides the
draw function.

Colors for pieces added to the constants file. if we don't like the
colors we can change them i was just putting something down.

for some of the functions there will need to be more conditions taken
into account. For example the fall function does not check to see if
there is a block under it.